requirements.txt
arrow==0.7.0
blessed==1.14.1
cycler==0.10.0
Django==1.9.2
django-getenv==1.3.1
django-model-utils==2.4
django-picklefield==0.3.2
django-q==0.7.15
djangorestframework==3.3.2
future==0.15.2
matplotlib==1.5.1
numpy==1.10.4
pep8==1.7.0
picamera==1.10
pyparsing==2.1.0
pyserial==3.0.1
python-dateutil==2.4.2
pytz==2015.7
RPi.GPIO==0.6.1
six==1.10.0
wcwidth==0.1.6
wheel==0.26.0
